+++ lib/mongoid-minitest/matchers/validations/format_of.rb
@@ -1,0 +1,9 @@
+module Mongoid
+  module Matchers
+    module Validations
+      def validate_format_of(field)
+        HaveValidationMatcher.new(field, :format)
+      end
+    end
+  end
+end