+++ test/models/person.rb
@@ -13,6 +13,6 @@ class Person
   validates_presence_of(:name)
   validates_uniqueness_of(:email)
   validates_uniqueness_of(:login, case_sensitive: false)
-
   validates_length_of(:password, minimum: 8, maximum: 16)
+  validates_format_of(:email, with: //)
 end